more cast casting to c++ casts

+++ src/LifeMeterBar.cpp
@@ -225,13 +225,13 @@ void LifeMeterBar::ChangeLife( float fDeltaLife )
 	}
 	else
 	{
-		fDeltaLife *= 1 + (float)m_iProgressiveLifebar/8 * m_iMissCombo;
+		fDeltaLife *= 1.f + static_cast<float>(m_iProgressiveLifebar)/8.f * static_cast<float>(m_iMissCombo);
 		// do this after; only successive W5/miss will increase the amount of life lost.
 		m_iMissCombo++;
 		/* Increase by m_iRegenComboAfterMiss; never push it beyond m_iMaxRegenComboAfterMiss
 		 * but don't reduce it if it's already past. */
 		const int NewComboToRegainLife = min(
-			 (int)PREFSMAN->m_iMaxRegenComboAfterMiss,
+			static_cast<int>(PREFSMAN->m_iMaxRegenComboAfterMiss),
 			 m_iComboToRegainLife + PREFSMAN->m_iRegenComboAfterMiss );
 
 		m_iComboToRegainLife = max( m_iComboToRegainLife, NewComboToRegainLife );
@@ -367,7 +367,7 @@ void LifeMeterBar::UpdateNonstopLifebar()
 	}
 
 	if( iTotal > 1 )
-		m_fLifeDifficulty = m_fBaseLifeDifficulty - 0.2f * (int)(iProgressiveLifebarDifficulty * iCleared / (iTotal - 1));
+		m_fLifeDifficulty = m_fBaseLifeDifficulty - 0.2f * static_cast<int>((iProgressiveLifebarDifficulty * iCleared / (iTotal - 1)));
 	else
 		m_fLifeDifficulty = m_fBaseLifeDifficulty - 0.2f * iProgressiveLifebarDifficulty;
